Network.AWS.CognitoSync

Description

Amazon Cognito Sync provides an AWS service and client library that enable cross-device syncing of application-related user data. High-level client libraries are available for both iOS and Android. You can use these libraries to persist data locally so that it's available even if the device is offline. Developer credentials don't need to be stored on the mobile device to access the service. You can use Amazon Cognito to obtain a normalized user ID and credentials. User data is persisted in a dataset that can store up to 1 MB of key-value pairs, and you can have up to 20 datasets per user identity.

Documentation

module Network.AWS.CognitoSync.BulkPublish

module Network.AWS.CognitoSync.DeleteDataset

module Network.AWS.CognitoSync.DescribeDataset

module Network.AWS.CognitoSync.DescribeIdentityPoolUsage

module Network.AWS.CognitoSync.DescribeIdentityUsage

module Network.AWS.CognitoSync.GetBulkPublishDetails

module Network.AWS.CognitoSync.GetIdentityPoolConfiguration

module Network.AWS.CognitoSync.ListDatasets

module Network.AWS.CognitoSync.ListIdentityPoolUsage

module Network.AWS.CognitoSync.ListRecords

module Network.AWS.CognitoSync.RegisterDevice

module Network.AWS.CognitoSync.SetIdentityPoolConfiguration

module Network.AWS.CognitoSync.SubscribeToDataset

module Network.AWS.CognitoSync.Types

module Network.AWS.CognitoSync.UnsubscribeFromDataset

module Network.AWS.CognitoSync.UpdateRecords