Network.AWS.CognitoSync.BulkPublish

Description

Initiates a bulk publish of all existing datasets for an Identity Pool to the configured stream. Customers are limited to one successful bulk publish per 24 hours. Bulk publish is an asynchronous request, customers can see the status of the request via the GetBulkPublishDetails operation.

http://docs.aws.amazon.com/cognitosync/latest/APIReference/API_BulkPublish.html

Request

data BulkPublish

Request constructor

bulkPublish

Arguments

:: Text
bpIdentityPoolId

-> BulkPublish

BulkPublish constructor.

The fields accessible through corresponding lenses are:

bpIdentityPoolId :: Text

Request lenses

bpIdentityPoolId :: Lens' BulkPublish Text

A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

Response

data BulkPublishResponse

Response constructor

bulkPublishResponse :: BulkPublishResponse

BulkPublishResponse constructor.

The fields accessible through corresponding lenses are:

bprIdentityPoolId :: Maybe Text

Response lenses

bprIdentityPoolId :: Lens' BulkPublishResponse (Maybe Text)

A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.