Network.AWS.CognitoSync.DescribeDataset

Description

Gets meta data about a dataset by identity and dataset name. With Amazon Cognito Sync, each identity has access only to its own data. Thus, the credentials used to make this API call need to have access to the identity data.

DescribeDataset can be called with temporary user credentials provided by Cognito Identity or with developer credentials. You should use Cognito Identity credentials to make this API call.

http://docs.aws.amazon.com/cognitosync/latest/APIReference/API_DescribeDataset.html

Request

data DescribeDataset

Request constructor

describeDataset

Arguments

:: Text
ddIdentityPoolId

-> Text
ddIdentityId

-> Text
ddDatasetName

-> DescribeDataset

DescribeDataset constructor.

The fields accessible through corresponding lenses are:

ddDatasetName :: Text

ddIdentityId :: Text

ddIdentityPoolId :: Text

Request lenses

ddDatasetName :: Lens' DescribeDataset Text

A string of up to 128 characters. Allowed characters are a-z, A-Z, 0-9, '_' (underscore), - (dash), and . (dot).

ddIdentityId :: Lens' DescribeDataset Text

A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

ddIdentityPoolId :: Lens' DescribeDataset Text

A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

Response

data DescribeDatasetResponse

Response constructor

describeDatasetResponse :: DescribeDatasetResponse

DescribeDatasetResponse constructor.

The fields accessible through corresponding lenses are:

ddrDataset :: Maybe Dataset

Response lenses

ddrDataset :: Lens' DescribeDatasetResponse (Maybe Dataset)

Meta data for a collection of data for an identity. An identity can have multiple datasets. A dataset can be general or associated with a particular entity in an application (like a saved game). Datasets are automatically created if they don't exist. Data is synced by dataset, and a dataset can hold up to 1MB of key-value pairs.