Network.AWS.CognitoSync.DescribeIdentityUsage

Description

Gets usage information for an identity, including number of datasets and data usage.

DescribeIdentityUsage can be called with temporary user credentials provided by Cognito Identity or with developer credentials.

http://docs.aws.amazon.com/cognitosync/latest/APIReference/API_DescribeIdentityUsage.html

Request

data DescribeIdentityUsage

Request constructor

describeIdentityUsage

Arguments

:: Text
diuIdentityPoolId

-> Text
diuIdentityId

-> DescribeIdentityUsage

DescribeIdentityUsage constructor.

The fields accessible through corresponding lenses are:

diuIdentityId :: Text

diuIdentityPoolId :: Text

Request lenses

diuIdentityId :: Lens' DescribeIdentityUsage Text

A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

diuIdentityPoolId :: Lens' DescribeIdentityUsage Text

A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

Response

data DescribeIdentityUsageResponse

Response constructor

describeIdentityUsageResponse :: DescribeIdentityUsageResponse

DescribeIdentityUsageResponse constructor.

The fields accessible through corresponding lenses are:

diurIdentityUsage :: Maybe IdentityUsage

Response lenses

diurIdentityUsage :: Lens' DescribeIdentityUsageResponse (Maybe IdentityUsage)

Usage information for the identity.