Network.AWS.CognitoSync.GetBulkPublishDetails

Description

Get the status of the last BulkPublish operation for an identity pool.

This API can only be called with developer credentials. You cannot call this API with the temporary user credentials provided by Cognito Identity.

Creating a Request

getBulkPublishDetails

Arguments

:: Text
gbpdIdentityPoolId

-> GetBulkPublishDetails

Creates a value of GetBulkPublishDetails with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gbpdIdentityPoolId - A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

data GetBulkPublishDetails

The input for the GetBulkPublishDetails operation.

See: getBulkPublishDetails smart constructor.

Request Lenses

gbpdIdentityPoolId :: Lens' GetBulkPublishDetails Text

A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

Destructuring the Response

getBulkPublishDetailsResponse

Arguments

:: Int
gbpdrsResponseStatus

-> GetBulkPublishDetailsResponse

Creates a value of GetBulkPublishDetailsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gbpdrsBulkPublishStartTime - The date/time at which the last bulk publish was initiated.

gbpdrsIdentityPoolId - A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

gbpdrsBulkPublishCompleteTime - If BulkPublishStatus is SUCCEEDED, the time the last bulk publish operation completed.

gbpdrsFailureMessage - If BulkPublishStatus is FAILED this field will contain the error message that caused the bulk publish to fail.

gbpdrsBulkPublishStatus - Status of the last bulk publish operation, valid values are: NOT_STARTED - No bulk publish has been requested for this identity pool IN_PROGRESS - Data is being published to the configured stream SUCCEEDED - All data for the identity pool has been published to the configured stream FAILED - Some portion of the data has failed to publish, check FailureMessage for the cause.

gbpdrsResponseStatus - -- | The response status code.

data GetBulkPublishDetailsResponse

The output for the GetBulkPublishDetails operation.

See: getBulkPublishDetailsResponse smart constructor.

Response Lenses

gbpdrsBulkPublishStartTime :: Lens' GetBulkPublishDetailsResponse (Maybe UTCTime)

The date/time at which the last bulk publish was initiated.

gbpdrsIdentityPoolId :: Lens' GetBulkPublishDetailsResponse (Maybe Text)

A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

gbpdrsBulkPublishCompleteTime :: Lens' GetBulkPublishDetailsResponse (Maybe UTCTime)

If BulkPublishStatus is SUCCEEDED, the time the last bulk publish operation completed.

gbpdrsFailureMessage :: Lens' GetBulkPublishDetailsResponse (Maybe Text)

If BulkPublishStatus is FAILED this field will contain the error message that caused the bulk publish to fail.

gbpdrsBulkPublishStatus :: Lens' GetBulkPublishDetailsResponse (Maybe BulkPublishStatus)

Status of the last bulk publish operation, valid values are: NOT_STARTED - No bulk publish has been requested for this identity pool IN_PROGRESS - Data is being published to the configured stream SUCCEEDED - All data for the identity pool has been published to the configured stream FAILED - Some portion of the data has failed to publish, check FailureMessage for the cause.

gbpdrsResponseStatus :: Lens' GetBulkPublishDetailsResponse Int

- | The response status code.