Network.AWS.CognitoSync.GetCognitoEvents

Description

Gets the events and the corresponding Lambda functions associated with an identity pool.

This API can only be called with developer credentials. You cannot call this API with the temporary user credentials provided by Cognito Identity.

Creating a Request

getCognitoEvents

Arguments

:: Text
gceIdentityPoolId

-> GetCognitoEvents

Creates a value of GetCognitoEvents with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gceIdentityPoolId - The Cognito Identity Pool ID for the request

data GetCognitoEvents

A request for a list of the configured Cognito Events

See: getCognitoEvents smart constructor.

Request Lenses

gceIdentityPoolId :: Lens' GetCognitoEvents Text

The Cognito Identity Pool ID for the request

Destructuring the Response

getCognitoEventsResponse

Arguments

:: Int
gcersResponseStatus

-> GetCognitoEventsResponse

Creates a value of GetCognitoEventsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gcersEvents - The Cognito Events returned from the GetCognitoEvents request

gcersResponseStatus - -- | The response status code.

data GetCognitoEventsResponse

The response from the GetCognitoEvents request

See: getCognitoEventsResponse smart constructor.

Response Lenses

gcersEvents :: Lens' GetCognitoEventsResponse (HashMap Text Text)

The Cognito Events returned from the GetCognitoEvents request

gcersResponseStatus :: Lens' GetCognitoEventsResponse Int

- | The response status code.