Network.AWS.CognitoSync.GetIdentityPoolConfiguration

Description

Gets the configuration settings of an identity pool.

http://docs.aws.amazon.com/cognitosync/latest/APIReference/API_GetIdentityPoolConfiguration.html

Request

data GetIdentityPoolConfiguration

Request constructor

getIdentityPoolConfiguration

Arguments

:: Text
gipcIdentityPoolId

-> GetIdentityPoolConfiguration

GetIdentityPoolConfiguration constructor.

The fields accessible through corresponding lenses are:

gipcIdentityPoolId :: Text

Request lenses

gipcIdentityPoolId :: Lens' GetIdentityPoolConfiguration Text

A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. This is the ID of the pool for which to return a configuration.

Response

data GetIdentityPoolConfigurationResponse

Response constructor

getIdentityPoolConfigurationResponse :: GetIdentityPoolConfigurationResponse

GetIdentityPoolConfigurationResponse constructor.

The fields accessible through corresponding lenses are:

gipcrCognitoStreams :: Maybe CognitoStreams

gipcrIdentityPoolId :: Maybe Text

gipcrPushSync :: Maybe PushSync

Response lenses

gipcrCognitoStreams :: Lens' GetIdentityPoolConfigurationResponse (Maybe CognitoStreams)

Options to apply to this identity pool for Amazon Cognito streams.

gipcrIdentityPoolId :: Lens' GetIdentityPoolConfigurationResponse (Maybe Text)

A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito.

gipcrPushSync :: Lens' GetIdentityPoolConfigurationResponse (Maybe PushSync)

Options to apply to this identity pool for push synchronization.