Network.AWS.CognitoSync.ListDatasets

Description

Lists datasets for an identity. With Amazon Cognito Sync, each identity has access only to its own data. Thus, the credentials used to make this API call need to have access to the identity data.

ListDatasets can be called with temporary user credentials provided by Cognito Identity or with developer credentials. You should use the Cognito Identity credentials to make this API call.

http://docs.aws.amazon.com/cognitosync/latest/APIReference/API_ListDatasets.html

Request

data ListDatasets

Request constructor

listDatasets

Arguments

:: Text
ldIdentityPoolId

-> Text
ldIdentityId

-> ListDatasets

ListDatasets constructor.

The fields accessible through corresponding lenses are:

ldIdentityId :: Text

ldIdentityPoolId :: Text

ldMaxResults :: Maybe Int

ldNextToken :: Maybe Text

Request lenses

ldIdentityId :: Lens' ListDatasets Text

A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

ldIdentityPoolId :: Lens' ListDatasets Text

A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

ldMaxResults :: Lens' ListDatasets (Maybe Int)

The maximum number of results to be returned.

ldNextToken :: Lens' ListDatasets (Maybe Text)

A pagination token for obtaining the next page of results.

Response

data ListDatasetsResponse

Response constructor

listDatasetsResponse :: ListDatasetsResponse

ListDatasetsResponse constructor.

The fields accessible through corresponding lenses are:

ldrCount :: Maybe Int

ldrDatasets :: [Dataset]

ldrNextToken :: Maybe Text

Response lenses

ldrCount :: Lens' ListDatasetsResponse (Maybe Int)

Number of datasets returned.

ldrDatasets :: Lens' ListDatasetsResponse [Dataset]

A set of datasets.

ldrNextToken :: Lens' ListDatasetsResponse (Maybe Text)

A pagination token for obtaining the next page of results.