Network.AWS.CognitoSync.ListIdentityPoolUsage

Description

Gets a list of identity pools registered with Cognito.

ListIdentityPoolUsage can only be called with developer credentials. You cannot make this API call with the temporary user credentials provided by Cognito Identity.

http://docs.aws.amazon.com/cognitosync/latest/APIReference/API_ListIdentityPoolUsage.html

Request

data ListIdentityPoolUsage

Request constructor

listIdentityPoolUsage :: ListIdentityPoolUsage

ListIdentityPoolUsage constructor.

The fields accessible through corresponding lenses are:

lipuMaxResults :: Maybe Int

lipuNextToken :: Maybe Text

Request lenses

lipuMaxResults :: Lens' ListIdentityPoolUsage (Maybe Int)

The maximum number of results to be returned.

lipuNextToken :: Lens' ListIdentityPoolUsage (Maybe Text)

A pagination token for obtaining the next page of results.

Response

data ListIdentityPoolUsageResponse

Response constructor

listIdentityPoolUsageResponse :: ListIdentityPoolUsageResponse

ListIdentityPoolUsageResponse constructor.

The fields accessible through corresponding lenses are:

lipurCount :: Maybe Int

lipurIdentityPoolUsages :: [IdentityPoolUsage]

lipurMaxResults :: Maybe Int

lipurNextToken :: Maybe Text

Response lenses

lipurCount :: Lens' ListIdentityPoolUsageResponse (Maybe Int)

Total number of identities for the identity pool.

lipurIdentityPoolUsages :: Lens' ListIdentityPoolUsageResponse [IdentityPoolUsage]

Usage information for the identity pools.

lipurMaxResults :: Lens' ListIdentityPoolUsageResponse (Maybe Int)

The maximum number of results to be returned.

lipurNextToken :: Lens' ListIdentityPoolUsageResponse (Maybe Text)

A pagination token for obtaining the next page of results.