Network.AWS.CognitoSync.ListRecords

Description

Gets paginated records, optionally changed after a particular sync count for a dataset and identity. With Amazon Cognito Sync, each identity has access only to its own data. Thus, the credentials used to make this API call need to have access to the identity data.

ListRecords can be called with temporary user credentials provided by Cognito Identity or with developer credentials. You should use Cognito Identity credentials to make this API call.

http://docs.aws.amazon.com/cognitosync/latest/APIReference/API_ListRecords.html

Request

data ListRecords

Request constructor

listRecords

Arguments

:: Text
lrIdentityPoolId

-> Text
lrIdentityId

-> Text
lrDatasetName

-> ListRecords

ListRecords constructor.

The fields accessible through corresponding lenses are:

lrDatasetName :: Text

lrIdentityId :: Text

lrIdentityPoolId :: Text

lrLastSyncCount :: Maybe Integer

lrMaxResults :: Maybe Int

lrNextToken :: Maybe Text

lrSyncSessionToken :: Maybe Text

Request lenses

lrDatasetName :: Lens' ListRecords Text

A string of up to 128 characters. Allowed characters are a-z, A-Z, 0-9, '_' (underscore), - (dash), and . (dot).

lrIdentityId :: Lens' ListRecords Text

A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

lrIdentityPoolId :: Lens' ListRecords Text

A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

lrLastSyncCount :: Lens' ListRecords (Maybe Integer)

The last server sync count for this record.

lrMaxResults :: Lens' ListRecords (Maybe Int)

The maximum number of results to be returned.

lrNextToken :: Lens' ListRecords (Maybe Text)

A pagination token for obtaining the next page of results.

lrSyncSessionToken :: Lens' ListRecords (Maybe Text)

A token containing a session ID, identity ID, and expiration.

Response

data ListRecordsResponse

Response constructor

listRecordsResponse :: ListRecordsResponse

ListRecordsResponse constructor.

The fields accessible through corresponding lenses are:

lrrCount :: Maybe Int

lrrDatasetDeletedAfterRequestedSyncCount :: Maybe Bool

lrrDatasetExists :: Maybe Bool

lrrDatasetSyncCount :: Maybe Integer

lrrLastModifiedBy :: Maybe Text

lrrMergedDatasetNames :: [Text]

lrrNextToken :: Maybe Text

lrrRecords :: [Record]

lrrSyncSessionToken :: Maybe Text

Response lenses

lrrCount :: Lens' ListRecordsResponse (Maybe Int)

Total number of records.

lrrDatasetDeletedAfterRequestedSyncCount :: Lens' ListRecordsResponse (Maybe Bool)

A boolean value specifying whether to delete the dataset locally.

lrrDatasetExists :: Lens' ListRecordsResponse (Maybe Bool)

Indicates whether the dataset exists.

lrrDatasetSyncCount :: Lens' ListRecordsResponse (Maybe Integer)

Server sync count for this dataset.

lrrLastModifiedBy :: Lens' ListRecordsResponse (Maybe Text)

The user/device that made the last change to this record.

lrrMergedDatasetNames :: Lens' ListRecordsResponse [Text]

Names of merged datasets.

lrrNextToken :: Lens' ListRecordsResponse (Maybe Text)

A pagination token for obtaining the next page of results.

lrrRecords :: Lens' ListRecordsResponse [Record]

A list of all records.

lrrSyncSessionToken :: Lens' ListRecordsResponse (Maybe Text)

A token containing a session ID, identity ID, and expiration.