Network.AWS.CognitoSync.RegisterDevice

Description

Registers a device to receive push sync notifications.

This API can only be called with temporary credentials provided by Cognito Identity. You cannot call this API with developer credentials.

Creating a Request

registerDevice

Arguments

:: Text
rdIdentityPoolId

-> Text
rdIdentityId

-> Platform
rdPlatform

-> Text
rdToken

-> RegisterDevice

Creates a value of RegisterDevice with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rdIdentityPoolId - A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. Here, the ID of the pool that the identity belongs to.

rdIdentityId - The unique ID for this identity.

rdPlatform - The SNS platform type (e.g. GCM, SDM, APNS, APNS_SANDBOX).

rdToken - The push token.

data RegisterDevice

A request to RegisterDevice.

See: registerDevice smart constructor.

Request Lenses

rdIdentityPoolId :: Lens' RegisterDevice Text

A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. Here, the ID of the pool that the identity belongs to.

rdIdentityId :: Lens' RegisterDevice Text

The unique ID for this identity.

rdPlatform :: Lens' RegisterDevice Platform

The SNS platform type (e.g. GCM, SDM, APNS, APNS_SANDBOX).

rdToken :: Lens' RegisterDevice Text

The push token.

Destructuring the Response

registerDeviceResponse

Arguments

:: Int
rdrsResponseStatus

-> RegisterDeviceResponse

Creates a value of RegisterDeviceResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rdrsDeviceId - The unique ID generated for this device by Cognito.

rdrsResponseStatus - -- | The response status code.

data RegisterDeviceResponse

Response to a RegisterDevice request.

See: registerDeviceResponse smart constructor.

Response Lenses

rdrsDeviceId :: Lens' RegisterDeviceResponse (Maybe Text)

The unique ID generated for this device by Cognito.

rdrsResponseStatus :: Lens' RegisterDeviceResponse Int

- | The response status code.