Network.AWS.CognitoSync.SetIdentityPoolConfiguration

Description

Sets the necessary configuration for push sync.

http://docs.aws.amazon.com/cognitosync/latest/APIReference/API_SetIdentityPoolConfiguration.html

Request

data SetIdentityPoolConfiguration

Request constructor

setIdentityPoolConfiguration

Arguments

:: Text
sipcIdentityPoolId

-> SetIdentityPoolConfiguration

SetIdentityPoolConfiguration constructor.

The fields accessible through corresponding lenses are:

sipcCognitoStreams :: Maybe CognitoStreams

sipcIdentityPoolId :: Text

sipcPushSync :: Maybe PushSync

Request lenses

sipcCognitoStreams :: Lens' SetIdentityPoolConfiguration (Maybe CognitoStreams)

Options to apply to this identity pool for Amazon Cognito streams.

sipcIdentityPoolId :: Lens' SetIdentityPoolConfiguration Text

A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. This is the ID of the pool to modify.

sipcPushSync :: Lens' SetIdentityPoolConfiguration (Maybe PushSync)

Options to apply to this identity pool for push synchronization.

Response

data SetIdentityPoolConfigurationResponse

Response constructor

setIdentityPoolConfigurationResponse :: SetIdentityPoolConfigurationResponse

SetIdentityPoolConfigurationResponse constructor.

The fields accessible through corresponding lenses are:

sipcrCognitoStreams :: Maybe CognitoStreams

sipcrIdentityPoolId :: Maybe Text

sipcrPushSync :: Maybe PushSync

Response lenses

sipcrCognitoStreams :: Lens' SetIdentityPoolConfigurationResponse (Maybe CognitoStreams)

Options to apply to this identity pool for Amazon Cognito streams.

sipcrIdentityPoolId :: Lens' SetIdentityPoolConfigurationResponse (Maybe Text)

A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito.

sipcrPushSync :: Lens' SetIdentityPoolConfigurationResponse (Maybe PushSync)

Options to apply to this identity pool for push synchronization.