Network.AWS.CognitoSync.SubscribeToDataset

Description

Subscribes to receive notifications when a dataset is modified by another device.

http://docs.aws.amazon.com/cognitosync/latest/APIReference/API_SubscribeToDataset.html

Request

data SubscribeToDataset

Request constructor

subscribeToDataset

Arguments

:: Text
stdIdentityPoolId

-> Text
stdIdentityId

-> Text
stdDatasetName

-> Text
stdDeviceId

-> SubscribeToDataset

SubscribeToDataset constructor.

The fields accessible through corresponding lenses are:

stdDatasetName :: Text

stdDeviceId :: Text

stdIdentityId :: Text

stdIdentityPoolId :: Text

Request lenses

stdDatasetName :: Lens' SubscribeToDataset Text

The name of the dataset to subcribe to.

stdDeviceId :: Lens' SubscribeToDataset Text

The unique ID generated for this device by Cognito.

stdIdentityId :: Lens' SubscribeToDataset Text

Unique ID for this identity.

stdIdentityPoolId :: Lens' SubscribeToDataset Text

A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. The ID of the pool to which the identity belongs.

Response

data SubscribeToDatasetResponse

Response constructor

subscribeToDatasetResponse :: SubscribeToDatasetResponse

SubscribeToDatasetResponse constructor.