Network.AWS.CognitoSync.Types

Service

data CognitoSync

Version 2014-06-30 of the Amazon Cognito Sync service.

Error

data JSONError :: *

IdentityPoolUsage

data IdentityPoolUsage

identityPoolUsage :: IdentityPoolUsage

IdentityPoolUsage constructor.

The fields accessible through corresponding lenses are:

ipuDataStorage :: Maybe Integer

ipuIdentityPoolId :: Maybe Text

ipuLastModifiedDate :: Maybe UTCTime

ipuSyncSessionsCount :: Maybe Integer

ipuDataStorage :: Lens' IdentityPoolUsage (Maybe Integer)

Data storage information for the identity pool.

ipuIdentityPoolId :: Lens' IdentityPoolUsage (Maybe Text)

A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

ipuLastModifiedDate :: Lens' IdentityPoolUsage (Maybe UTCTime)

Date on which the identity pool was last modified.

ipuSyncSessionsCount :: Lens' IdentityPoolUsage (Maybe Integer)

Number of sync sessions for the identity pool.

Platform

data Platform

Constructors

Adm
ADM

Apns
APNS

ApnsSandbox
APNS_SANDBOX

Gcm
GCM

Dataset

data Dataset

dataset :: Dataset

Dataset constructor.

The fields accessible through corresponding lenses are:

dCreationDate :: Maybe UTCTime

dDataStorage :: Maybe Integer

dDatasetName :: Maybe Text

dIdentityId :: Maybe Text

dLastModifiedBy :: Maybe Text

dLastModifiedDate :: Maybe UTCTime

dNumRecords :: Maybe Integer

dCreationDate :: Lens' Dataset (Maybe UTCTime)

Date on which the dataset was created.

dDataStorage :: Lens' Dataset (Maybe Integer)

Total size in bytes of the records in this dataset.

dDatasetName :: Lens' Dataset (Maybe Text)

A string of up to 128 characters. Allowed characters are a-z, A-Z, 0-9, '_' (underscore), - (dash), and . (dot).

dIdentityId :: Lens' Dataset (Maybe Text)

A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

dLastModifiedBy :: Lens' Dataset (Maybe Text)

The device that made the last change to this dataset.

dLastModifiedDate :: Lens' Dataset (Maybe UTCTime)

Date when the dataset was last modified.

dNumRecords :: Lens' Dataset (Maybe Integer)

Number of records in this dataset.

Operation

data Operation

Constructors

Remove
remove

Replace
replace

StreamingStatus

data StreamingStatus

Constructors

Disabled
DISABLED

Enabled
ENABLED

BulkPublishStatus

data BulkPublishStatus

Constructors

Failed
FAILED

InProgress
IN_PROGRESS

NotStarted
NOT_STARTED

Succeeded
SUCCEEDED

Record

data Record

record :: Record

Record constructor.

The fields accessible through corresponding lenses are:

rDeviceLastModifiedDate :: Maybe UTCTime

rKey :: Maybe Text

rLastModifiedBy :: Maybe Text

rLastModifiedDate :: Maybe UTCTime

rSyncCount :: Maybe Integer

rValue :: Maybe Text

rDeviceLastModifiedDate :: Lens' Record (Maybe UTCTime)

The last modified date of the client device.

rKey :: Lens' Record (Maybe Text)

The key for the record.

rLastModifiedBy :: Lens' Record (Maybe Text)

The user/device that made the last change to this record.

rLastModifiedDate :: Lens' Record (Maybe UTCTime)

The date on which the record was last modified.

rSyncCount :: Lens' Record (Maybe Integer)

The server sync count for this record.

rValue :: Lens' Record (Maybe Text)

The value for the record.

CognitoStreams

data CognitoStreams

cognitoStreams :: CognitoStreams

CognitoStreams constructor.

The fields accessible through corresponding lenses are:

csRoleArn :: Maybe Text

csStreamName :: Maybe Text

csStreamingStatus :: Maybe StreamingStatus

csRoleArn :: Lens' CognitoStreams (Maybe Text)

The ARN of the role Amazon Cognito can assume in order to publish to the stream. This role must grant access to Amazon Cognito (cognito-sync) to invoke PutRecord on your Cognito stream.

csStreamName :: Lens' CognitoStreams (Maybe Text)

The name of the Cognito stream to receive updates. This stream must be in the developers account and in the same region as the identity pool.

csStreamingStatus :: Lens' CognitoStreams (Maybe StreamingStatus)

Status of the Cognito streams. Valid values are: ENABLED - Streaming of updates to identity pool is enabled.

DISABLEDStreaming of updates to identity pool is disabled. Bulk publish will also fail if StreamingStatus is DISABLED.

IdentityUsage

data IdentityUsage

identityUsage :: IdentityUsage

IdentityUsage constructor.

The fields accessible through corresponding lenses are:

iuDataStorage :: Maybe Integer

iuDatasetCount :: Maybe Int

iuIdentityId :: Maybe Text

iuIdentityPoolId :: Maybe Text

iuLastModifiedDate :: Maybe UTCTime

iuDataStorage :: Lens' IdentityUsage (Maybe Integer)

Total data storage for this identity.

iuDatasetCount :: Lens' IdentityUsage (Maybe Int)

Number of datasets for the identity.

iuIdentityId :: Lens' IdentityUsage (Maybe Text)

A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

iuIdentityPoolId :: Lens' IdentityUsage (Maybe Text)

A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

iuLastModifiedDate :: Lens' IdentityUsage (Maybe UTCTime)

Date on which the identity was last modified.

RecordPatch

data RecordPatch

recordPatch

Arguments

:: Operation
rpOp

-> Text
rpKey

-> Integer
rpSyncCount

-> RecordPatch

RecordPatch constructor.

The fields accessible through corresponding lenses are:

rpDeviceLastModifiedDate :: Maybe UTCTime

rpKey :: Text

rpOp :: Operation

rpSyncCount :: Integer

rpValue :: Maybe Text

rpDeviceLastModifiedDate :: Lens' RecordPatch (Maybe UTCTime)

The last modified date of the client device.

rpKey :: Lens' RecordPatch Text

The key associated with the record patch.

rpOp :: Lens' RecordPatch Operation

An operation, either replace or remove.

rpSyncCount :: Lens' RecordPatch Integer

Last known server sync count for this record. Set to 0 if unknown.

rpValue :: Lens' RecordPatch (Maybe Text)

The value associated with the record patch.

PushSync

data PushSync

pushSync :: PushSync

PushSync constructor.

The fields accessible through corresponding lenses are:

psApplicationArns :: [Text]

psRoleArn :: Maybe Text

psApplicationArns :: Lens' PushSync [Text]

List of SNS platform application ARNs that could be used by clients.

psRoleArn :: Lens' PushSync (Maybe Text)

A role configured to allow Cognito to call SNS on behalf of the developer.