Network.AWS.CognitoSync.UnsubscribeFromDataset

Description

Unsubscribes from receiving notifications when a dataset is modified by another device.

This API can only be called with temporary credentials provided by Cognito Identity. You cannot call this API with developer credentials.

Creating a Request

unsubscribeFromDataset

Arguments

:: Text
ufdIdentityPoolId

-> Text
ufdIdentityId

-> Text
ufdDatasetName

-> Text
ufdDeviceId

-> UnsubscribeFromDataset

Creates a value of UnsubscribeFromDataset with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ufdIdentityPoolId - A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. The ID of the pool to which this identity belongs.

ufdIdentityId - Unique ID for this identity.

ufdDatasetName - The name of the dataset from which to unsubcribe.

ufdDeviceId - The unique ID generated for this device by Cognito.

data UnsubscribeFromDataset

A request to UnsubscribeFromDataset.

See: unsubscribeFromDataset smart constructor.

Request Lenses

ufdIdentityPoolId :: Lens' UnsubscribeFromDataset Text

A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. The ID of the pool to which this identity belongs.

ufdIdentityId :: Lens' UnsubscribeFromDataset Text

Unique ID for this identity.

ufdDatasetName :: Lens' UnsubscribeFromDataset Text

The name of the dataset from which to unsubcribe.

ufdDeviceId :: Lens' UnsubscribeFromDataset Text

The unique ID generated for this device by Cognito.

Destructuring the Response

unsubscribeFromDatasetResponse

Arguments

:: Int
ufdrsResponseStatus

-> UnsubscribeFromDatasetResponse

Creates a value of UnsubscribeFromDatasetResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ufdrsResponseStatus - -- | The response status code.

data UnsubscribeFromDatasetResponse

Response to an UnsubscribeFromDataset request.

See: unsubscribeFromDatasetResponse smart constructor.

Response Lenses

ufdrsResponseStatus :: Lens' UnsubscribeFromDatasetResponse Int

- | The response status code.