Network.AWS.CognitoSync.UpdateRecords

Description

Posts updates to records and adds and deletes records for a dataset and user.

UpdateRecords can only be called with temporary user credentials provided by Cognito Identity. You cannot make this API call with developer credentials.

http://docs.aws.amazon.com/cognitosync/latest/APIReference/API_UpdateRecords.html

Request

data UpdateRecords

Request constructor

updateRecords

Arguments

:: Text
urIdentityPoolId

-> Text
urIdentityId

-> Text
urDatasetName

-> Text
urSyncSessionToken

-> UpdateRecords

UpdateRecords constructor.

The fields accessible through corresponding lenses are:

urClientContext :: Maybe Text

urDatasetName :: Text

urDeviceId :: Maybe Text

urIdentityId :: Text

urIdentityPoolId :: Text

urRecordPatches :: [RecordPatch]

urSyncSessionToken :: Text

Request lenses

urClientContext :: Lens' UpdateRecords (Maybe Text)

Intended to supply a device ID that will populate the lastModifiedBy field referenced in other methods. The ClientContext field is not yet implemented.

urDatasetName :: Lens' UpdateRecords Text

A string of up to 128 characters. Allowed characters are a-z, A-Z, 0-9, '_' (underscore), - (dash), and . (dot).

urDeviceId :: Lens' UpdateRecords (Maybe Text)

The unique ID generated for this device by Cognito.

urIdentityId :: Lens' UpdateRecords Text

A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

urIdentityPoolId :: Lens' UpdateRecords Text

A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

urRecordPatches :: Lens' UpdateRecords [RecordPatch]

A list of patch operations.

urSyncSessionToken :: Lens' UpdateRecords Text

The SyncSessionToken returned by a previous call to ListRecords for this dataset and identity.

Response

data UpdateRecordsResponse

Response constructor

updateRecordsResponse :: UpdateRecordsResponse

UpdateRecordsResponse constructor.

The fields accessible through corresponding lenses are:

urrRecords :: [Record]

Response lenses

urrRecords :: Lens' UpdateRecordsResponse [Record]

A list of records that have been updated.