Network.AWS.CognitoSync

Description

Amazon Cognito Sync

Amazon Cognito Sync provides an AWS service and client library that enable cross-device syncing of application-related user data. High-level client libraries are available for both iOS and Android. You can use these libraries to persist data locally so that it's available even if the device is offline. Developer credentials don't need to be stored on the mobile device to access the service. You can use Amazon Cognito to obtain a normalized user ID and credentials. User data is persisted in a dataset that can store up to 1 MB of key-value pairs, and you can have up to 20 datasets per user identity.

With Amazon Cognito Sync, the data stored for each identity is accessible only to credentials assigned to that identity. In order to use the Cognito Sync service, you need to make API calls using credentials retrieved with Amazon Cognito Identity service .

If you want to use Cognito Sync in an Android or iOS application, you will probably want to make API calls via the AWS Mobile SDK. To learn more, see the Developer Guide for Android and the Developer Guide for iOS .

Service Configuration

cognitoSync :: Service

API version 2014-06-30 of the Amazon Cognito Sync SDK configuration.

Errors

Error matchers are designed for use with the functions provided by Control.Exception.Lens. This allows catching (and rethrowing) service specific errors returned by CognitoSync.

InvalidParameterException

_InvalidParameterException :: AsError a => Getting (First ServiceError) a ServiceError

Thrown when a request parameter does not comply with the associated constraints.

NotAuthorizedException

_NotAuthorizedException :: AsError a => Getting (First ServiceError) a ServiceError

Thrown when a user is not authorized to access the requested resource.

InternalErrorException

_InternalErrorException :: AsError a => Getting (First ServiceError) a ServiceError

Indicates an internal service error.

InvalidConfigurationException

_InvalidConfigurationException :: AsError a => Getting (First ServiceError) a ServiceError

Prism for InvalidConfigurationException' errors.

DuplicateRequestException

_DuplicateRequestException :: AsError a => Getting (First ServiceError) a ServiceError

An exception thrown when there is an IN_PROGRESS bulk publish operation for the given identity pool.

LambdaThrottledException

_LambdaThrottledException :: AsError a => Getting (First ServiceError) a ServiceError

AWS Lambda throttled your account, please contact AWS Support

AlreadyStreamedException

_AlreadyStreamedException :: AsError a => Getting (First ServiceError) a ServiceError

An exception thrown when a bulk publish operation is requested less than 24 hours after a previous bulk publish operation completed successfully.

InvalidLambdaFunctionOutputException

_InvalidLambdaFunctionOutputException :: AsError a => Getting (First ServiceError) a ServiceError

The AWS Lambda function returned invalid output or an exception.

TooManyRequestsException

_TooManyRequestsException :: AsError a => Getting (First ServiceError) a ServiceError

Thrown if the request is throttled.

ConcurrentModificationException

_ConcurrentModificationException :: AsError a => Getting (First ServiceError) a ServiceError

Thrown if there are parallel requests to modify a resource.

ResourceConflictException

_ResourceConflictException :: AsError a => Getting (First ServiceError) a ServiceError

Thrown if an update can't be applied because the resource was changed by another call and this would result in a conflict.

ResourceNotFoundException

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

Thrown if the resource doesn't exist.

LimitExceededException

_LimitExceededException :: AsError a => Getting (First ServiceError) a ServiceError

Thrown when the limit on the number of objects or operations has been exceeded.

Waiters

Waiters poll by repeatedly sending a request until some remote success condition configured by the Wait specification is fulfilled. The Wait specification determines how many attempts should be made, in addition to delay and retry strategies.

Operations

Some AWS operations return results that are incomplete and require subsequent requests in order to obtain the entire result set. The process of sending subsequent requests to continue where a previous request left off is called pagination. For example, the ListObjects operation of Amazon S3 returns up to 1000 objects at a time, and you must send subsequent requests with the appropriate Marker in order to retrieve the next page of results.

Operations that have an AWSPager instance can transparently perform subsequent requests, correctly setting Markers and other request facets to iterate through the entire result set of a truncated API operation. Operations which support this have an additional note in the documentation.

Many operations have the ability to filter results on the server side. See the individual operation parameters for details.

DescribeDataset

module Network.AWS.CognitoSync.DescribeDataset

SetCognitoEvents

module Network.AWS.CognitoSync.SetCognitoEvents

DescribeIdentityPoolUsage

module Network.AWS.CognitoSync.DescribeIdentityPoolUsage

GetBulkPublishDetails

module Network.AWS.CognitoSync.GetBulkPublishDetails

ListIdentityPoolUsage

module Network.AWS.CognitoSync.ListIdentityPoolUsage

SetIdentityPoolConfiguration

module Network.AWS.CognitoSync.SetIdentityPoolConfiguration

DeleteDataset

module Network.AWS.CognitoSync.DeleteDataset

GetCognitoEvents

module Network.AWS.CognitoSync.GetCognitoEvents

DescribeIdentityUsage

module Network.AWS.CognitoSync.DescribeIdentityUsage

RegisterDevice

module Network.AWS.CognitoSync.RegisterDevice

SubscribeToDataset

module Network.AWS.CognitoSync.SubscribeToDataset

GetIdentityPoolConfiguration

module Network.AWS.CognitoSync.GetIdentityPoolConfiguration

ListRecords

module Network.AWS.CognitoSync.ListRecords

UnsubscribeFromDataset

module Network.AWS.CognitoSync.UnsubscribeFromDataset

UpdateRecords

module Network.AWS.CognitoSync.UpdateRecords

ListDatasets

module Network.AWS.CognitoSync.ListDatasets

BulkPublish

module Network.AWS.CognitoSync.BulkPublish

Types

BulkPublishStatus

data BulkPublishStatus

Constructors

Failed

InProgress

NotStarted

Succeeded

Operation

data Operation

Constructors

Remove

Replace

Platform

data Platform

Constructors

ADM

APNS

APNSSandbox

GCM

StreamingStatus

data StreamingStatus

Constructors

Disabled

Enabled

CognitoStreams

data CognitoStreams

Configuration options for configure Cognito streams.

See: cognitoStreams smart constructor.

cognitoStreams :: CognitoStreams

Creates a value of CognitoStreams with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

csStreamingStatus - Status of the Cognito streams. Valid values are: ENABLED - Streaming of updates to identity pool is enabled. DISABLED - Streaming of updates to identity pool is disabled. Bulk publish will also fail if StreamingStatus is DISABLED.

csStreamName - The name of the Cognito stream to receive updates. This stream must be in the developers account and in the same region as the identity pool.

csRoleARN - The ARN of the role Amazon Cognito can assume in order to publish to the stream. This role must grant access to Amazon Cognito (cognito-sync) to invoke PutRecord on your Cognito stream.

csStreamingStatus :: Lens' CognitoStreams (Maybe StreamingStatus)

Status of the Cognito streams. Valid values are: ENABLED - Streaming of updates to identity pool is enabled. DISABLED - Streaming of updates to identity pool is disabled. Bulk publish will also fail if StreamingStatus is DISABLED.

csStreamName :: Lens' CognitoStreams (Maybe Text)

The name of the Cognito stream to receive updates. This stream must be in the developers account and in the same region as the identity pool.

csRoleARN :: Lens' CognitoStreams (Maybe Text)

The ARN of the role Amazon Cognito can assume in order to publish to the stream. This role must grant access to Amazon Cognito (cognito-sync) to invoke PutRecord on your Cognito stream.

Dataset

data Dataset

A collection of data for an identity pool. An identity pool can have multiple datasets. A dataset is per identity and can be general or associated with a particular entity in an application (like a saved game). Datasets are automatically created if they don't exist. Data is synced by dataset, and a dataset can hold up to 1MB of key-value pairs.

See: dataset smart constructor.

dataset :: Dataset

Creates a value of Dataset with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dLastModifiedDate - Date when the dataset was last modified.

dNumRecords - Number of records in this dataset.

dDataStorage - Total size in bytes of the records in this dataset.

dDatasetName - A string of up to 128 characters. Allowed characters are a-z, A-Z, 0-9, '_' (underscore), - (dash), and . (dot).

dCreationDate - Date on which the dataset was created.

dLastModifiedBy - The device that made the last change to this dataset.

dIdentityId - A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

dLastModifiedDate :: Lens' Dataset (Maybe UTCTime)

Date when the dataset was last modified.

dNumRecords :: Lens' Dataset (Maybe Integer)

Number of records in this dataset.

dDataStorage :: Lens' Dataset (Maybe Integer)

Total size in bytes of the records in this dataset.

dDatasetName :: Lens' Dataset (Maybe Text)

A string of up to 128 characters. Allowed characters are a-z, A-Z, 0-9, '_' (underscore), - (dash), and . (dot).

dCreationDate :: Lens' Dataset (Maybe UTCTime)

Date on which the dataset was created.

dLastModifiedBy :: Lens' Dataset (Maybe Text)

The device that made the last change to this dataset.

dIdentityId :: Lens' Dataset (Maybe Text)

A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

IdentityPoolUsage

data IdentityPoolUsage

Usage information for the identity pool.

See: identityPoolUsage smart constructor.

identityPoolUsage :: IdentityPoolUsage

Creates a value of IdentityPoolUsage with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ipuLastModifiedDate - Date on which the identity pool was last modified.

ipuIdentityPoolId - A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

ipuDataStorage - Data storage information for the identity pool.

ipuSyncSessionsCount - Number of sync sessions for the identity pool.

ipuLastModifiedDate :: Lens' IdentityPoolUsage (Maybe UTCTime)

Date on which the identity pool was last modified.

ipuIdentityPoolId :: Lens' IdentityPoolUsage (Maybe Text)

A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

ipuDataStorage :: Lens' IdentityPoolUsage (Maybe Integer)

Data storage information for the identity pool.

ipuSyncSessionsCount :: Lens' IdentityPoolUsage (Maybe Integer)

Number of sync sessions for the identity pool.

IdentityUsage

data IdentityUsage

Usage information for the identity.

See: identityUsage smart constructor.

identityUsage :: IdentityUsage

Creates a value of IdentityUsage with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

iuLastModifiedDate - Date on which the identity was last modified.

iuIdentityPoolId - A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

iuDatasetCount - Number of datasets for the identity.

iuDataStorage - Total data storage for this identity.

iuIdentityId - A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

iuLastModifiedDate :: Lens' IdentityUsage (Maybe UTCTime)

Date on which the identity was last modified.

iuIdentityPoolId :: Lens' IdentityUsage (Maybe Text)

A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

iuDatasetCount :: Lens' IdentityUsage (Maybe Int)

Number of datasets for the identity.

iuDataStorage :: Lens' IdentityUsage (Maybe Integer)

Total data storage for this identity.

iuIdentityId :: Lens' IdentityUsage (Maybe Text)

A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

PushSync

data PushSync

Configuration options to be applied to the identity pool.

See: pushSync smart constructor.

pushSync :: PushSync

Creates a value of PushSync with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

psApplicationARNs - List of SNS platform application ARNs that could be used by clients.

psRoleARN - A role configured to allow Cognito to call SNS on behalf of the developer.

psApplicationARNs :: Lens' PushSync [Text]

List of SNS platform application ARNs that could be used by clients.

psRoleARN :: Lens' PushSync (Maybe Text)

A role configured to allow Cognito to call SNS on behalf of the developer.

Record

data Record

The basic data structure of a dataset.

See: record smart constructor.

record :: Record

Creates a value of Record with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rSyncCount - The server sync count for this record.

rDeviceLastModifiedDate - The last modified date of the client device.

rLastModifiedDate - The date on which the record was last modified.

rValue - The value for the record.

rKey - The key for the record.

rLastModifiedBy - The user/device that made the last change to this record.

rSyncCount :: Lens' Record (Maybe Integer)

The server sync count for this record.

rDeviceLastModifiedDate :: Lens' Record (Maybe UTCTime)

The last modified date of the client device.

rLastModifiedDate :: Lens' Record (Maybe UTCTime)

The date on which the record was last modified.

rValue :: Lens' Record (Maybe Text)

The value for the record.

rKey :: Lens' Record (Maybe Text)

The key for the record.

rLastModifiedBy :: Lens' Record (Maybe Text)

The user/device that made the last change to this record.

RecordPatch

data RecordPatch

An update operation for a record.

See: recordPatch smart constructor.

recordPatch

Arguments

:: Operation
rpOp

-> Text
rpKey

-> Integer
rpSyncCount

-> RecordPatch

Creates a value of RecordPatch with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rpDeviceLastModifiedDate - The last modified date of the client device.

rpValue - The value associated with the record patch.

rpOp - An operation, either replace or remove.

rpKey - The key associated with the record patch.

rpSyncCount - Last known server sync count for this record. Set to 0 if unknown.

rpDeviceLastModifiedDate :: Lens' RecordPatch (Maybe UTCTime)

The last modified date of the client device.

rpValue :: Lens' RecordPatch (Maybe Text)

The value associated with the record patch.

rpOp :: Lens' RecordPatch Operation

An operation, either replace or remove.

rpKey :: Lens' RecordPatch Text

The key associated with the record patch.

rpSyncCount :: Lens' RecordPatch Integer

Last known server sync count for this record. Set to 0 if unknown.