Network.AWS.CognitoSync.DescribeDataset

Description

Gets meta data about a dataset by identity and dataset name. With Amazon Cognito Sync, each identity has access only to its own data. Thus, the credentials used to make this API call need to have access to the identity data.

This API can be called with temporary user credentials provided by Cognito Identity or with developer credentials. You should use Cognito Identity credentials to make this API call.

Creating a Request

describeDataset

Arguments

:: Text
ddIdentityPoolId

-> Text
ddIdentityId

-> Text
ddDatasetName

-> DescribeDataset

Creates a value of DescribeDataset with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ddIdentityPoolId - A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

ddIdentityId - A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

ddDatasetName - A string of up to 128 characters. Allowed characters are a-z, A-Z, 0-9, '_' (underscore), - (dash), and . (dot).

data DescribeDataset

A request for meta data about a dataset (creation date, number of records, size) by owner and dataset name.

See: describeDataset smart constructor.

Request Lenses

ddIdentityPoolId :: Lens' DescribeDataset Text

A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

ddIdentityId :: Lens' DescribeDataset Text

A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

ddDatasetName :: Lens' DescribeDataset Text

A string of up to 128 characters. Allowed characters are a-z, A-Z, 0-9, '_' (underscore), - (dash), and . (dot).

Destructuring the Response

describeDatasetResponse

Arguments

:: Int
ddrsResponseStatus

-> DescribeDatasetResponse

Creates a value of DescribeDatasetResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ddrsDataset - Meta data for a collection of data for an identity. An identity can have multiple datasets. A dataset can be general or associated with a particular entity in an application (like a saved game). Datasets are automatically created if they don't exist. Data is synced by dataset, and a dataset can hold up to 1MB of key-value pairs.

ddrsResponseStatus - -- | The response status code.

data DescribeDatasetResponse

Response to a successful DescribeDataset request.

See: describeDatasetResponse smart constructor.

Response Lenses

ddrsDataset :: Lens' DescribeDatasetResponse (Maybe Dataset)

Meta data for a collection of data for an identity. An identity can have multiple datasets. A dataset can be general or associated with a particular entity in an application (like a saved game). Datasets are automatically created if they don't exist. Data is synced by dataset, and a dataset can hold up to 1MB of key-value pairs.

ddrsResponseStatus :: Lens' DescribeDatasetResponse Int

- | The response status code.