Network.AWS.CognitoSync.DescribeIdentityUsage

Description

Gets usage information for an identity, including number of datasets and data usage.

This API can be called with temporary user credentials provided by Cognito Identity or with developer credentials.

Creating a Request

describeIdentityUsage

Arguments

:: Text
diuIdentityPoolId

-> Text
diuIdentityId

-> DescribeIdentityUsage

Creates a value of DescribeIdentityUsage with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

diuIdentityPoolId - A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

diuIdentityId - A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

data DescribeIdentityUsage

A request for information about the usage of an identity pool.

See: describeIdentityUsage smart constructor.

Request Lenses

diuIdentityPoolId :: Lens' DescribeIdentityUsage Text

A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

diuIdentityId :: Lens' DescribeIdentityUsage Text

A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

Destructuring the Response

describeIdentityUsageResponse

Arguments

:: Int
diursResponseStatus

-> DescribeIdentityUsageResponse

Creates a value of DescribeIdentityUsageResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

diursIdentityUsage - Usage information for the identity.

diursResponseStatus - -- | The response status code.

data DescribeIdentityUsageResponse

The response to a successful DescribeIdentityUsage request.

See: describeIdentityUsageResponse smart constructor.

Response Lenses

diursIdentityUsage :: Lens' DescribeIdentityUsageResponse (Maybe IdentityUsage)

Usage information for the identity.

diursResponseStatus :: Lens' DescribeIdentityUsageResponse Int

- | The response status code.