Network.AWS.CognitoSync.ListDatasets

Description

Lists datasets for an identity. With Amazon Cognito Sync, each identity has access only to its own data. Thus, the credentials used to make this API call need to have access to the identity data.

ListDatasets can be called with temporary user credentials provided by Cognito Identity or with developer credentials. You should use the Cognito Identity credentials to make this API call.

Creating a Request

listDatasets

Arguments

:: Text
ldIdentityId

-> Text
ldIdentityPoolId

-> ListDatasets

Creates a value of ListDatasets with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ldNextToken - A pagination token for obtaining the next page of results.

ldMaxResults - The maximum number of results to be returned.

ldIdentityId - A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

ldIdentityPoolId - A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

data ListDatasets

Request for a list of datasets for an identity.

See: listDatasets smart constructor.

Request Lenses

ldNextToken :: Lens' ListDatasets (Maybe Text)

A pagination token for obtaining the next page of results.

ldMaxResults :: Lens' ListDatasets (Maybe Int)

The maximum number of results to be returned.

ldIdentityId :: Lens' ListDatasets Text

A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

ldIdentityPoolId :: Lens' ListDatasets Text

A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

Destructuring the Response

listDatasetsResponse

Arguments

:: Int
ldrsResponseStatus

-> ListDatasetsResponse

Creates a value of ListDatasetsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ldrsCount - Number of datasets returned.

ldrsNextToken - A pagination token for obtaining the next page of results.

ldrsDatasets - A set of datasets.

ldrsResponseStatus - -- | The response status code.

data ListDatasetsResponse

Returned for a successful ListDatasets request.

See: listDatasetsResponse smart constructor.

Response Lenses

ldrsCount :: Lens' ListDatasetsResponse (Maybe Int)

Number of datasets returned.

ldrsNextToken :: Lens' ListDatasetsResponse (Maybe Text)

A pagination token for obtaining the next page of results.

ldrsDatasets :: Lens' ListDatasetsResponse [Dataset]

A set of datasets.

ldrsResponseStatus :: Lens' ListDatasetsResponse Int

- | The response status code.