Network.AWS.CognitoSync.ListIdentityPoolUsage

Description

Gets a list of identity pools registered with Cognito.

ListIdentityPoolUsage can only be called with developer credentials. You cannot make this API call with the temporary user credentials provided by Cognito Identity.

Creating a Request

listIdentityPoolUsage :: ListIdentityPoolUsage

Creates a value of ListIdentityPoolUsage with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lipuNextToken - A pagination token for obtaining the next page of results.

lipuMaxResults - The maximum number of results to be returned.

data ListIdentityPoolUsage

A request for usage information on an identity pool.

See: listIdentityPoolUsage smart constructor.

Request Lenses

lipuNextToken :: Lens' ListIdentityPoolUsage (Maybe Text)

A pagination token for obtaining the next page of results.

lipuMaxResults :: Lens' ListIdentityPoolUsage (Maybe Int)

The maximum number of results to be returned.

Destructuring the Response

listIdentityPoolUsageResponse

Arguments

:: Int
lipursResponseStatus

-> ListIdentityPoolUsageResponse

Creates a value of ListIdentityPoolUsageResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lipursIdentityPoolUsages - Usage information for the identity pools.

lipursCount - Total number of identities for the identity pool.

lipursNextToken - A pagination token for obtaining the next page of results.

lipursMaxResults - The maximum number of results to be returned.

lipursResponseStatus - -- | The response status code.

data ListIdentityPoolUsageResponse

Returned for a successful ListIdentityPoolUsage request.

See: listIdentityPoolUsageResponse smart constructor.

Response Lenses

lipursIdentityPoolUsages :: Lens' ListIdentityPoolUsageResponse [IdentityPoolUsage]

Usage information for the identity pools.

lipursCount :: Lens' ListIdentityPoolUsageResponse (Maybe Int)

Total number of identities for the identity pool.

lipursNextToken :: Lens' ListIdentityPoolUsageResponse (Maybe Text)

A pagination token for obtaining the next page of results.

lipursMaxResults :: Lens' ListIdentityPoolUsageResponse (Maybe Int)

The maximum number of results to be returned.

lipursResponseStatus :: Lens' ListIdentityPoolUsageResponse Int

- | The response status code.