Network.AWS.CognitoSync.ListRecords

Description

Gets paginated records, optionally changed after a particular sync count for a dataset and identity. With Amazon Cognito Sync, each identity has access only to its own data. Thus, the credentials used to make this API call need to have access to the identity data.

ListRecords can be called with temporary user credentials provided by Cognito Identity or with developer credentials. You should use Cognito Identity credentials to make this API call.

Creating a Request

listRecords

Arguments

:: Text
lrIdentityPoolId

-> Text
lrIdentityId

-> Text
lrDatasetName

-> ListRecords

Creates a value of ListRecords with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lrLastSyncCount - The last server sync count for this record.

lrNextToken - A pagination token for obtaining the next page of results.

lrSyncSessionToken - A token containing a session ID, identity ID, and expiration.

lrMaxResults - The maximum number of results to be returned.

lrIdentityPoolId - A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

lrIdentityId - A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

lrDatasetName - A string of up to 128 characters. Allowed characters are a-z, A-Z, 0-9, '_' (underscore), - (dash), and . (dot).

data ListRecords

A request for a list of records.

See: listRecords smart constructor.

Request Lenses

lrLastSyncCount :: Lens' ListRecords (Maybe Integer)

The last server sync count for this record.

lrNextToken :: Lens' ListRecords (Maybe Text)

A pagination token for obtaining the next page of results.

lrSyncSessionToken :: Lens' ListRecords (Maybe Text)

A token containing a session ID, identity ID, and expiration.

lrMaxResults :: Lens' ListRecords (Maybe Int)

The maximum number of results to be returned.

lrIdentityPoolId :: Lens' ListRecords Text

A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

lrIdentityId :: Lens' ListRecords Text

A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

lrDatasetName :: Lens' ListRecords Text

A string of up to 128 characters. Allowed characters are a-z, A-Z, 0-9, '_' (underscore), - (dash), and . (dot).

Destructuring the Response

listRecordsResponse

Arguments

:: Int
lrrsResponseStatus

-> ListRecordsResponse

Creates a value of ListRecordsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lrrsDatasetDeletedAfterRequestedSyncCount - A boolean value specifying whether to delete the dataset locally.

lrrsDatasetExists - Indicates whether the dataset exists.

lrrsCount - Total number of records.

lrrsRecords - A list of all records.

lrrsNextToken - A pagination token for obtaining the next page of results.

lrrsMergedDatasetNames - Names of merged datasets.

lrrsSyncSessionToken - A token containing a session ID, identity ID, and expiration.

lrrsLastModifiedBy - The user/device that made the last change to this record.

lrrsDatasetSyncCount - Server sync count for this dataset.

lrrsResponseStatus - -- | The response status code.

data ListRecordsResponse

Returned for a successful ListRecordsRequest.

See: listRecordsResponse smart constructor.

Response Lenses

lrrsDatasetDeletedAfterRequestedSyncCount :: Lens' ListRecordsResponse (Maybe Bool)

A boolean value specifying whether to delete the dataset locally.

lrrsDatasetExists :: Lens' ListRecordsResponse (Maybe Bool)

Indicates whether the dataset exists.

lrrsCount :: Lens' ListRecordsResponse (Maybe Int)

Total number of records.

lrrsRecords :: Lens' ListRecordsResponse [Record]

A list of all records.

lrrsNextToken :: Lens' ListRecordsResponse (Maybe Text)

A pagination token for obtaining the next page of results.

lrrsMergedDatasetNames :: Lens' ListRecordsResponse [Text]

Names of merged datasets.

lrrsSyncSessionToken :: Lens' ListRecordsResponse (Maybe Text)

A token containing a session ID, identity ID, and expiration.

lrrsLastModifiedBy :: Lens' ListRecordsResponse (Maybe Text)

The user/device that made the last change to this record.

lrrsDatasetSyncCount :: Lens' ListRecordsResponse (Maybe Integer)

Server sync count for this dataset.

lrrsResponseStatus :: Lens' ListRecordsResponse Int

- | The response status code.