Network.AWS.CognitoSync.UpdateRecords

Description

Posts updates to records and adds and deletes records for a dataset and user.

The sync count in the record patch is your last known sync count for that record. The server will reject an UpdateRecords request with a ResourceConflictException if you try to patch a record with a new value but a stale sync count.

For example, if the sync count on the server is 5 for a key called highScore and you try and submit a new highScore with sync count of 4, the request will be rejected. To obtain the current sync count for a record, call ListRecords. On a successful update of the record, the response returns the new sync count for that record. You should present that sync count the next time you try to update that same record. When the record does not exist, specify the sync count as 0.

This API can be called with temporary user credentials provided by Cognito Identity or with developer credentials.

Creating a Request

updateRecords

Arguments

:: Text
urIdentityPoolId

-> Text
urIdentityId

-> Text
urDatasetName

-> Text
urSyncSessionToken

-> UpdateRecords

Creates a value of UpdateRecords with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

urRecordPatches - A list of patch operations.

urDeviceId - The unique ID generated for this device by Cognito.

urClientContext - Intended to supply a device ID that will populate the lastModifiedBy field referenced in other methods. The ClientContext field is not yet implemented.

urIdentityPoolId - A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

urIdentityId - A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

urDatasetName - A string of up to 128 characters. Allowed characters are a-z, A-Z, 0-9, '_' (underscore), - (dash), and . (dot).

urSyncSessionToken - The SyncSessionToken returned by a previous call to ListRecords for this dataset and identity.

data UpdateRecords

A request to post updates to records or add and delete records for a dataset and user.

See: updateRecords smart constructor.

Request Lenses

urRecordPatches :: Lens' UpdateRecords [RecordPatch]

A list of patch operations.

urDeviceId :: Lens' UpdateRecords (Maybe Text)

The unique ID generated for this device by Cognito.

urClientContext :: Lens' UpdateRecords (Maybe Text)

Intended to supply a device ID that will populate the lastModifiedBy field referenced in other methods. The ClientContext field is not yet implemented.

urIdentityPoolId :: Lens' UpdateRecords Text

A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

urIdentityId :: Lens' UpdateRecords Text

A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.

urDatasetName :: Lens' UpdateRecords Text

A string of up to 128 characters. Allowed characters are a-z, A-Z, 0-9, '_' (underscore), - (dash), and . (dot).

urSyncSessionToken :: Lens' UpdateRecords Text

The SyncSessionToken returned by a previous call to ListRecords for this dataset and identity.

Destructuring the Response

updateRecordsResponse

Arguments

:: Int
urrsResponseStatus

-> UpdateRecordsResponse

Creates a value of UpdateRecordsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

urrsRecords - A list of records that have been updated.

urrsResponseStatus - -- | The response status code.

data UpdateRecordsResponse

Returned for a successful UpdateRecordsRequest.

See: updateRecordsResponse smart constructor.

Response Lenses

urrsRecords :: Lens' UpdateRecordsResponse [Record]

A list of records that have been updated.

urrsResponseStatus :: Lens' UpdateRecordsResponse Int

- | The response status code.